employment_portal/requirements.txt
Babel
backports.shutil-get-terminal-size
boto
boto3
botocore
decorator
defusedxml
django-allauth
django-autofixture
django-bootstrap3
django-countries
django-extensions
django-localflavor
django-phonenumber-field
django-storages
docutils
httplib2
ipython
ipython-genutils
jmespath
oauth2
oauthlib
pexpect
phonenumberslite
pickleshare
Pillow
ptyprocess
PyJWT
python-dateutil
python-http-client
python-openid
python3-openid
pytz
requests
requests-oauthlib
sendgrid
sendgrid-django
shortuuid
simplegeneric
six
smtpapi
traitlets